feat: Transforms + DT +  Asset loader working

+++ Engine/src/Architecture/ECS/ECS_Test.cpp
@@ -9,7 +9,7 @@ namespace IonixEngine {
     void SpriteRenderer::Render(RenderData* data) {
         Vec2 position = entity->transform.GetGlobalPosition();
         Vec2 scale = entity->transform.GetGlobalScale();
-        SDL_Rect draw = { static_cast<int>(position.x - (50 * scale.x)), static_cast<int>(position.y - (50 * scale.y)), 100 * scale.x, 100 * scale.y };
+        SDL_Rect draw = { static_cast<int>(position.x - (50 * scale.x)), static_cast<int>(position.y - (50 * scale.y)), static_cast<int>(100 * scale.x), static_cast<int>(100 * scale.y) };
 
         if (SDL_RenderCopy(data->renderer, image, nullptr, &draw) < 0) {
             SDL_Log("Could not draw texture: %s", SDL_GetError());
